requirements_minimal.txt
streamlit==1.29.0
plotly==5.18.0
pandas==2.2.0
numpy==1.26.0
requests==2.28.1
pydantic==1.10.12
langchain==0.0.170
langchain-google-genai==1.0.10
python-dotenv==0.20.0
beautifulsoup4==4.11.1
scraperapi-sdk==1.5.3
tiktoken==0.3.2
python-decouple==3.6
lxml==4.9.2
